# docs/swagger.yaml
basePath: /v1
definitions:
...
  contract.ReqUserActivated:
    properties:
      activation_token:
        maxLength: 26
        minLength: 26
        type: string
    required:
    - activation_token
    type: object
  contract.ReqUserDeposit:
    properties:
      amount:
        minimum: 50000
        type: number
    required:
    - amount
    type: object
  contract.ReqUserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  contract.ReqUserReactivate:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  contract.ReqUserRegister:
    properties:
      address:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - address
    - email
    - first_name
    - last_name
    - password
    - phone
    type: object
...
  contract.ResResendActivationToken:
    properties:
      message:
        type: string
    type: object
  contract.ResUserActivated:
    properties:
      user:
        properties:
          activated:
            type: boolean
          address:
            type: string
          deposit:
            type: number
          email:
            type: string
          first_name:
            type: string
          last_name:
            type: string
          phone:
            type: string
          user_id:
            type: integer
        required:
        - address
        - phone
        type: object
    type: object
  contract.ResUserDeposit:
    properties:
      message:
        type: string
    type: object
  contract.ResUserLogin:
    properties:
      auhentication_token:
        properties:
          expiry:
            type: string
          token:
            type: string
        type: object
    type: object
  contract.ResUserRegister:
    properties:
      user:
        properties:
          activated:
            type: boolean
          address:
            type: string
          deposit:
            type: number
          email:
            type: string
          first_name:
            type: string
          last_name:
            type: string
          phone:
            type: string
          user_id:
            type: integer
        required:
        - address
        - phone
        type: object
    type: object
host: localhost:8080
info:
  contact:
    email: ucokkocu411@gmail.com
    name: ucok-man
  description: Documentation for PixelRental API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PixelRental API
  version: "1.0"
paths:
...
  /users/activated:
    post:
      consumes:
      - application/json
      description: Activate registered user record
      parameters:
      - description: Activation Token
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/contract.ReqUserActivated'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResUserActivated'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Activate user record
      tags:
      - users
  /users/deposit:
    post:
      consumes:
      - application/json
      description: Top up saldo user record
      parameters:
      - description: Deposit Saldo
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/contract.ReqUserDeposit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResUserDeposit'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Deposit saldo user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Login user record
      parameters:
      - description: Login User
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/contract.ReqUserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResUserLogin'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Login user
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Get info of login user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResUserDeposit'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Get profile user
      tags:
      - users
  /users/reactivated:
    post:
      consumes:
      - application/json
      description: Resending new activation token user
      parameters:
      - description: Email User
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/contract.ReqUserReactivate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.ResResendActivationToken'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Resend activation token
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Create new user record
      parameters:
      - description: Create User
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/contract.ReqUserRegister'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/contract.ResUserRegister'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "429":
          description: Too Many Requests
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                properties:
                  message:
                    type: string
                type: object
            type: object
      summary: Create user
      tags:
      - users
swagger: "2.0"